Session_6/mailer/pyproject.toml
[project]
name = "mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "docx>=0.2.4",
    "mcp>=1.11.0",
    "google-genai>=1.25.0",
    "python-docx>=1.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
]

[tool.poetry.dependencies]
requests = "*"
beautifulsoup4 = "*"
python-docx = "*"
rich = "*"
